feat: US-015 - Add Route Metadata and Loading States
