+++ frontend/src/app/predictions/loading.tsx
@@ -1,0 +1,101 @@
+import { PageLayout } from '@/components/layout/PageLayout';
+import { Card, CardContent, CardHeader } from '@/components/ui/card';
+import { Skeleton } from '@/components/ui/skeleton';
+
+export default function PredictionsLoading() {
+  return (
+    <PageLayout>
+      {/* Page Header Skeleton */}
+      <div className="mb-8">
+        <Skeleton className="mb-2 h-9 w-64" />
+        <Skeleton className="h-5 w-96" />
+      </div>
+
+      {/* Tournament Status Card Skeleton */}
+      <Card className="mb-8">
+        <CardContent className="flex flex-col gap-4 py-4 sm:flex-row sm:items-center sm:justify-between">
+          <div className="flex items-center gap-4">
+            <div className="flex items-center gap-2">
+              <Skeleton className="h-5 w-5" />
+              <div>
+                <Skeleton className="mb-1 h-4 w-16" />
+                <Skeleton className="h-5 w-24" />
+              </div>
+            </div>
+            <Skeleton className="h-6 w-32" />
+          </div>
+          <div className="flex items-center gap-2">
+            <Skeleton className="h-5 w-5" />
+            <div>
+              <Skeleton className="mb-1 h-4 w-16" />
+              <Skeleton className="h-5 w-20" />
+            </div>
+          </div>
+        </CardContent>
+      </Card>
+
+      {/* Validation Summary Card Skeleton */}
+      <Card className="mb-8">
+        <CardHeader className="pb-3">
+          <Skeleton className="mb-1 h-5 w-40" />
+          <Skeleton className="h-4 w-64" />
+        </CardHeader>
+        <CardContent>
+          <div className="grid gap-4 sm:grid-cols-3">
+            {Array.from({ length: 3 }).map((_, i) => (
+              <div key={i} className="flex items-center gap-3">
+                <Skeleton className="h-5 w-5" />
+                <div>
+                  <Skeleton className="mb-1 h-4 w-24" />
+                  <Skeleton className="h-3 w-20" />
+                </div>
+              </div>
+            ))}
+          </div>
+        </CardContent>
+      </Card>
+
+      {/* Tabs Skeleton */}
+      <div className="mb-8">
+        <div className="mb-6 grid w-full grid-cols-3 gap-1">
+          {Array.from({ length: 3 }).map((_, i) => (
+            <Skeleton key={i} className="h-10 w-full" />
+          ))}
+        </div>
+
+        {/* Groups Grid Skeleton */}
+        <div className="grid gap-4 sm:grid-cols-2 lg:grid-cols-3">
+          {Array.from({ length: 6 }).map((_, i) => (
+            <Card key={i}>
+              <CardHeader className="pb-3">
+                <div className="flex items-center justify-between">
+                  <Skeleton className="h-5 w-20" />
+                  <Skeleton className="h-5 w-16" />
+                </div>
+              </CardHeader>
+              <CardContent className="space-y-2">
+                {Array.from({ length: 4 }).map((_, j) => (
+                  <div key={j} className="flex items-center gap-2">
+                    <Skeleton className="h-4 w-8" />
+                    <Skeleton className="h-9 w-full" />
+                  </div>
+                ))}
+              </CardContent>
+            </Card>
+          ))}
+        </div>
+      </div>
+
+      {/* Submit Section Skeleton */}
+      <Card className="border-primary/20 bg-primary/5">
+        <CardContent className="flex flex-col items-center gap-4 py-6 sm:flex-row sm:justify-between">
+          <div>
+            <Skeleton className="mb-1 h-5 w-32" />
+            <Skeleton className="h-4 w-56" />
+          </div>
+          <Skeleton className="h-11 w-44" />
+        </CardContent>
+      </Card>
+    </PageLayout>
+  );
+}
